==> 01_preprocessing/02_reject_incomplete_dates.ipynb <==
import pandas as pd
from datetime import datetime

# read relevant data
file = '../DATA/01_observations_decoded.csv'
df = pd.read_csv(file, delimiter=';')

# group data by timestamp
df_groups = df.groupby('timestamp')

# function to filter days with low data volume
def filter(groups, observations):
    result = []
    cc = 1
    l = len(groups)
    pc = 1
    for date, data in groups:
        # start compute progress
        perc = (cc/l)*100
        if perc > pc:
            print('{0}: {1:.1f}%'.format(datetime.now(), perc))
            pc += 1
        cc += 1
        
        # filter low-vaulme data
        if len(data) >= observations:
            for idx, row in data.iterrows():
                result.append([row['name'], row['east'], row['north'], row['up'], row['timestamp'], row['value']])
    return result

# run function, create and export dataframe as csv
data = filter(df_groups, 129)
result = pd.DataFrame(data=data, columns=df.columns)
result.sort_values(['timestamp', 'name'])
result.to_csv('../DATA/02_observations_high-data-volume.csv', sep=';', index=False)
