==> 01_preprocessing/03_reject_redundant_zero_values.ipynb <==
import pandas as pd
from datetime import datetime

# read relevant data
file_in = '../DATA/02_observations_high-data-volume.csv'
df = pd.read_csv(file_in, delimiter=';')

# group data by timestamp
groups = df.groupby('timestamp')

# function to check, if there where only zero-values for a day
def only_zero_data(dataframe):
    for row in dataframe.iterrows():
        value = row[1]['value']
        if value != 0:
            return False
    return True

# run function for each group(day), write to non-redundant groups to csv-file
file_out = '../DATA/03_valid_observations.csv'
once = False
header = True

l = len(groups)
pc = 1
cc = 1

for name, group in groups:
    # compute progress
    perc = (cc/l)*100
    if perc > pc:
        print('{0}: {1:.1f}%'.format(datetime.now(), perc))
        pc+=1
    cc+=1
    group.sort_values(['timestamp', 'name'])
    current_df = only_zero_data(group)
    if once == False:
        if current_df == True:
            if header:
                group.to_csv(file_out, sep=';', index=False, mode='a')
                once = True
                header = False
            else:
                group.to_csv(file_out, sep=';', header=False, index=False, mode='a')
                once = True
    if current_df == False:
        if header:
            group.to_csv(file_out, sep=';', index=False, mode='a')
            header = False
        else:
            group.to_csv(file_out, sep=';', header=False, index=False, mode='a')
