==> 01_preprocessing/05_replace_nearest_station-names_by_values.ipynb <==
import pandas as pd
from datetime import datetime
from pathlib import Path

# read relevant data - create temporal file with new order
temp_file = '../DATA/05_temp.csv'
temp = Path(temp_file)
if temp.is_file() == False:
    file = '../DATA/04_station_nearest_stations-names.csv'
    df = pd.read_csv(file, delimiter=';', encoding='latin1')
    df = df.sort_values(['timestamp', 'name'])
    df.to_csv('../DATA/05_temp.csv', sep=';', index=False)

# read in re-orderd temporary file
with open ('../DATA/05_temp.csv', 'r', encoding='utf-8') as file:
    lines = file.readlines()
lines = [x.strip() for x in lines]
header = lines[0]
del lines[0]

# replace station-names by their values measured at each day
n = 50 # number of neighbors to replace
cc = 1
pc = 1
le = len(lines)/129
with open ('../DATA/05_station_nearest_stations.csv', 'a') as file_out:
    file_out.write('{}\n'.format(header))
    for date in range(int(len(lines)/129)):
        # compute progress
        perc = (cc/le)*100
        if perc > pc:
            print('{0}: {1:.1f}%'.format(datetime.now(), perc))
            pc+=1
        cc+=1
        # replace names with values
        l = int(date*129)
        h = int((date+1)*129)
        data = lines[l:h]
        d = {}
        for line in data:    # prepare dict with name-value pairs
            split=line.split(';')
            d[split[0]] = split[5]
            
        for line in data:   # replace stationsnames with values for each line
            split=line.split(';')
            for i in range(n):
                name = split[7+(2*i)]
                split[7+(2*i)] = d[name]
            cr = ';'.join(str(e) for e in split)
            file_out.write('{}\n'.format(cr))
