==> 01_preprocessing/07_prepare_training-dataset.ipynb <==
import pandas as pd
from dateutil.parser import parse
from datetime import datetime

# read relevant data
obs_file = '../DATA/05_station_nearest_stations.csv'
obs = pd.read_csv(obs_file, delimiter=';', encoding='latin')

# reformat old timestamp to a ML-valid representation
result = []
l = len(obs)
pc = 1
cc = 0
for idx, row in obs.iterrows():
    # compute progress
    perc = (cc/l)*100
    if perc > pc:
        print('{0}: {1:.1f}%'.format(datetime.now(), pc))
        pc += 1
    cc += 1
    # reformat timestamp to doy and unixtime
    time = datetime.strptime(row['timestamp'], '%Y-%m-%d')
    unix = int((time-datetime(1970,1,1)).total_seconds())
    doy = time.timetuple().tm_yday
    result.append([unix, doy])

# create dataframe for new time-representations and concatenate with origin data
time_df = pd.DataFrame(result, columns=['unix', 'doy'])
df = pd.concat([obs, time_df], axis=1, join='outer')

# drop old timestamp
df = df.drop('timestamp', axis=1)

# export to csv
file = '../DATA/07_training_data.csv'
df.to_csv(file, sep=';', index=False)
