==> 01_preprocessing/08_prepare_prediction-dataset.ipynb <==
import pandas as pd
import datetime

# read relevant data
pred_file = '../DATA/06_dgm_nearest_station-names.csv'
pred = pd.read_csv(pred_file, delimiter=';', encoding='ansi')

obs_file = '../DATA/07_day_of_observation.csv' # manually extracted sub-file from 07_training_data.csv
obs_df = pd.read_csv(obs_file, delimiter=';')

# dictionary with name-value-pairs
data = {}
for row in obs_df.iterrows():
    cr = row[1]
    data[cr['name']] = cr['value']

# function that replaces the name with their values
def replace_names_by_values(df, data, unix, doy, number=3):
    cc = 1
    pc = 1
    l = len(pred)
    
    for idx, row in df.iterrows():
        # compute progress
        perc = (cc/l)*100
        if perc > pc:
            print('{0}: {1:.1f}%'.format(datetime.datetime.now(), pc))
            pc = pc+1
        cc = cc+1
        
        # replace names with values from dictionary
        df.at[idx, 'unix'] = unix
        df.at[idx, 'doy'] = doy
        for n in range(number):
            df.at[idx, 'value_{}'.format(n+1)] = data[row['value_{}'.format(n+1)]]    
    return df

# run function to replace names by values
pred = replace_names_by_values(pred, data, obs_df['unix'][0], obs_df['doy'][0], 50)

# sort data by north, east
pred = pred.sort_values(['north', 'east'], ascending=[False, True])

# export as csv
file = '../DATA/07_points_of_prediction.csv'
pred.to_csv(file, sep=';', index=False)
